#3 added the finalize seeds and first endpoint

+++ server/models/customer.js
@@ -14,7 +14,7 @@ module.exports = (sequelize, DataTypes) => {
     has_ordered: DataTypes.BOOLEAN,
     has_newsletter: DataTypes.BOOLEAN,
     nb_commands: DataTypes.INTEGER,
-    total_spent: DataTypes.INTEGER
+    total_spent: DataTypes.DECIMAL
   }, {
     tableName: 'customer',
     underscored: true,
